[ANDROSDK-640] Refactor in package structure
